update import according to restructure, bump version to 0.3.0
