+++ setup.cfg
@@ -9,7 +9,7 @@ long_description = file: README.md
 long_description_content_type = text/markdown
 license = BSD
 license_file = LICENSE
-keywords = python, bayes, naive bayes, classifier, probabilistic
+keywords = python, machine learning, bayes, naive bayes, classifier, probabilistic
 classifiers =
     Intended Audience :: Science/Research
     Intended Audience :: Developers
