+++ wnb/__init__.py
@@ -2,14 +2,14 @@
 Python library for the implementations of general and weighted naive Bayes (WNB) classifiers.
 """
 
-__version__ = "0.2.7"
+__version__ = "0.3.0"
 __author__ = "Mehdi Samsami"
 
 
-from .base import ContinuousDistMixin, DiscreteDistMixin
-from .enums import Distribution
-from .gnb import GeneralNB
-from .gwnb import GaussianWNB
+from wnb.gnb import GeneralNB
+from wnb.gwnb import GaussianWNB
+from wnb.stats.base import ContinuousDistMixin, DiscreteDistMixin
+from wnb.stats.enums import Distribution
 
 __all__ = [
     "GeneralNB",
